Adding qbp directory
